% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appmodels.R
\name{appmodels}
\alias{appmodels}
\alias{models}
\alias{machine}
\alias{learning}
\alias{soft}
\alias{voting}
\alias{classes}
\title{appmodels}
\usage{
appmodels(mtx, m, n, len, cls)
}
\arguments{
\item{mtx}{data.frame of mn-mers dataset}

\item{m}{number of the mnmer}

\item{n}{number of the mnmer}

\item{len}{character of the length of the fragment (550, 1000, 3000, 5000, 10000 bp)}

\item{cls}{character represents which class is applied class1 or class2}
}
\value{
dataframe
}
\description{
This function convert to a mnmer matrix and apply machine learning models
}
\keyword{classification}
\keyword{learning}
\keyword{machine}
\keyword{models}
\keyword{softvoting}
