% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sequences.R
\name{predict_sequences}
\alias{predict_sequences}
\alias{arbovirus}
\alias{mosquito-specific}
\alias{classification}
\title{predict_sequences}
\usage{
predict_sequences(seqs, cutoff = 0.5, all.data = TRUE)
}
\arguments{
\item{seqs}{DNAStringSet object to perform the classification}

\item{cutoff}{is the probability limit for the soft voting of the models in ML}

\item{all.data}{logical indicates if all sequence must be shown or only mosquito-specific and arbovirus}
}
\value{
dataframe
}
\description{
This function classifies sequences for the groups of arbovirus, mosquito-specific or othervirus
}
\examples{
 
seqfile <- system.file ('extdata/PRJNA778885_known.fasta.gz',package="MosVir")
seqs <- readDNAStringSet(seqfile)
res <- predict_sequences (seqs, cutoff=0.5, all.data=TRUE)

}
\keyword{arbovirus}
\keyword{classification}
\keyword{mosquito}
