% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fasta.R
\name{save_fasta}
\alias{save_fasta}
\alias{save}
\alias{fasta}
\alias{format}
\alias{results}
\alias{soft}
\alias{voting}
\alias{classes}
\title{save_fasta}
\usage{
save_fasta(seqs, res, file, category = "all")
}
\arguments{
\item{seqs}{DNAStringSet object for the FASTA sequences}

\item{res}{data.frame with the results of the predict_sequences function}

\item{file}{string for the output file to write a FASTA file}

\item{group}{string of the groups: all (default), mosquito, mosquito-specific or arboviruses}
}
\description{
This function save a FASTA file of the predictions results
}
\keyword{classification}
\keyword{learning}
\keyword{machine}
\keyword{models}
\keyword{softvoting}
